Fixed small error in signup invalidating certain passwords. Refactored some files.
