+++ src/main/java/com/example/projet2/Main.java
@@ -2,13 +2,11 @@ package com.example.projet2;
 
 import com.example.projet2.database.DatabaseManager;
 import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
 import javafx.stage.Stage;
 
 import java.io.IOException;
 
-public class HelloApplication extends Application {
+public class Main extends Application {
     @Override
     public void start(Stage stage) throws IOException {
         DatabaseManager.initializeDatabase();
