+++ src/main/resources/com/example/projet2/dashboard-view.fxml
@@ -7,7 +7,7 @@
 
 <VBox xmlns="http://javafx.com/javafx"
       xmlns:fx="http://javafx.com/fxml"
-      fx:controller="com.example.projet2.DashboardController"
+      fx:controller="com.example.projet2.sceneControllers.DashboardController"
       spacing="20"
       alignment="CENTER"
       prefWidth="600"
