+++ src/main/resources/com/example/projet2/transaction-view.fxml
@@ -8,7 +8,7 @@
 
 <VBox xmlns="http://javafx.com/javafx"
       xmlns:fx="http://javafx.com/fxml"
-      fx:controller="com.example.projet2.TransactionController"
+      fx:controller="com.example.projet2.sceneControllers.TransactionController"
       spacing="15"
       alignment="CENTER"
       prefWidth="600"
